+++ CHANGELOG.md
@@ -48,6 +48,8 @@ and this project adheres to
 -   Drop support of `ia_code`, in favor of always inserting the correct one.
 -   Deprecate `AAC_SAMPLE_FREQUENCY_INDEX_ESCAPE_VALUE` and explicit
     `sampling_frequency`, these are usually forbidden by the IAMF spec.
+-   Deprecate `AmbisonicsMode` enum and `AmbisonicsConfig::ambisonics_mode`
+    field. Infer the type based on which `ambisonics_config` is present.
 -   Deprecate the spelling of `HEADPHONES_RENDERING_MODE_BINAURAL` in favor of
     `HEADPHONES_RENDERING_MODE_BINAURAL_WORLD_LOCKED`.
 -   Deprecate the spelling of `HEADPHONES_RENDERING_MODE_RESERVED_2` in favor of